requirements.txt
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.7
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonify==0.5
MarkupSafe==2.1.5
oauthlib==3.2.2
proto-plus==1.24.0
protobuf==5.27.3
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.0.3
zipp==3.19.2
